vote and order projection fixed

+++ Booklandbackend/src/main/java/com/bookland/demobookland/model/projections/OrderSimpleProjection.java
@@ -12,5 +12,5 @@ public interface OrderSimpleProjection {
 
     String getAddressTitle();
 
-    Integer getTotalAmount();
+    Float getTotalAmount();
 }
